minor tweaks before release
